Renamed RubikSolver.java to clear confusion
